+++ dwds/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Remove dependency on `package:build_daemon`.
 - Add `FrontendServerAssetReader` for use with Frontend Server builds.
 - Depend on latest `package:sse` for handling client reconnects transparently on the server.
+- Fix an issue where trying to launch DevTools in a non-debug enabled Chrome
+  instance could crash your development server.
 
 **Breaking Changes:**
 - No longer use the `BuildResult` abstraction from `package:build_daemon` but